requirements.txt
bcrypt==3.1.7
cffi==1.12.3
easydict==1.9
motor==2.0.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
simplejson==3.16.0
six==1.12.0
tornado==6.0.3
